Update Task Progress Database
- Changed marks from integer to float to support decimals

+++ database/migrations/2018_10_04_172649_create_tasks_progress_table.php
@@ -19,7 +19,7 @@ class CreateTasksProgressTable extends Migration
             $table->integer('lab_id');
             $table->integer('task_id');
             $table->boolean('status');
-            $table->integer('marks');
+            $table->float('marks');
             $table->timestamps();
             $table->softDeletes();
         });
